# docs/openapi/api-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebDotNetMentoringProgram",
    "version": "1.0"
  },
  "paths": {
...
    "/api/CategoriesApi/GetImageById": {
      "get": {
        "tags": [
          "GetImageById"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CategoriesApi/UpdateImage": {
      "post": {
        "tags": [
          "UpdateImage"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "image",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}